Separate code from headers, add some addresses for testing on private server and clean up some code
