+++ pointers.txt
@@ -1,0 +1,14 @@
+CharName 00C79D18
+Rotation 00BEBA70
+Health 00000000
+MaxHealth 00000000
+Mana 00000000
+MaxMana 00000000
+Experience 00000000
+Alive 00000000
+Targetting 00000000
+TarName 00000000
+TarHealth 00000000
+TarHealthMax 00000000
+TarMana 00000000
+TarManaMax 00000000
