+++ src/ecmWaypoint.h
@@ -37,8 +37,6 @@ struct waypoint
     }
 };
 
-
-
 inline float magSq(const waypoint& w)
 {
     return w.x * w.x + w.y * w.y + w.z * w.z;
@@ -54,23 +52,6 @@ inline float rotationBetween(const waypoint& w1, const waypoint& w2)
     return atan2f(w1.x - w2.x, w1.y - w2.y);
 }
 
-inline int closest(const waypoint& pos, vector<waypoint> *pWaypoints)
-{
-    if (pWaypoints->size() < 1)
-        return -1;
-
-    float minDistance = FLT_MAX;
-    int minIndex = 0;
-    for (unsigned int i = 0; i < pWaypoints->size(); ++i)
-    {
-        float dist = magSq(pWaypoints->at(i) - pos);
-        if (dist < minDistance)
-        {
-            minIndex = i;
-            minDistance = dist;
-        }
-    }
-    return minIndex;
-}
+int closest(const waypoint& pos, vector<waypoint>* pWaypoints);
 
 #endif
